+++ SteelCutOptimizer.Server/Structs/DTO/CuttingStockResultsDTO.cs
@@ -4,9 +4,9 @@ namespace SteelCutOptimizer.Server.DTO
 {
     public class CuttingStockResultsDTO
     {
-        public int? TotalCost { get; set; }
-        public int TotalWaste { get; set; }
-        public int? TotalRelax { get; set; }
+        public long? TotalCost { get; set; }
+        public long TotalWaste { get; set; }
+        public long? TotalRelax { get; set; }
         public List<ResultPattern> Patterns { get; set; } = [];
         public List<double> OrderPrices { get; set; } = [];
         public List<double> StockLimits { get; set; } = [];
